Move profiler support into resources
