+++ resources/profilers/__init__.py
@@ -1,0 +1,1 @@
+"""Bundled profiler support resources."""
